requirements.txt
fastapi
uvicorn
sqlmodel
mysqlclient
kafka-python
boto3